+++ core/src/main/java/com/netflix/conductor/core/execution/archival/ExternalServiceConfiguration.java
@@ -22,7 +22,7 @@ public class ExternalServiceConfiguration {
 
     @Singleton
     public ApiProperties getReminderEndpoint() {
-        return new ApiProperties(configuration.getProperty("reminder.host", "reminder-service.swiggy.prod"),
+        return new ApiProperties(configuration.getProperty("reminder.host", "http://reminder-service.swiggy.prod"),
                 Integer.parseInt(configuration.getProperty("reminder.connect.timeout", "300")),
                 Integer.parseInt(configuration.getProperty("reminder.read.timeout", "300")));
     }
